[URLFollow] split out Steam
