+++ desertbot/modules/urlfollow/Twitter.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 """
-Created on Mar 13, 2014
+Created on Jan 25, 2014
 
 @author: StarlitGhost
 """
